Extracted base spec from HttpDownloaderSpec
